+++ packages/react-scripts/bin/react-scripts.js
@@ -19,16 +19,16 @@ const spawn = require('react-dev-utils/crossSpawn');
 const args = process.argv.slice(2);
 
 const scriptIndex = args.findIndex(
-  x => x === 'build' || x === 'eject' || x === 'start' || x === 'test'
+  x => x === 'build' || x === 'build-ssr' || x === 'eject' || x === 'start' || x === 'start-ssr' || x === 'test'
 );
 const script = scriptIndex === -1 ? args[0] : args[scriptIndex];
 const nodeArgs = scriptIndex > 0 ? args.slice(0, scriptIndex) : [];
 
-if (['build', 'eject', 'start', 'test'].includes(script)) {
+if (['build', 'build-ssr', 'eject', 'start', 'start-ssr', 'test'].includes(script)) {
   const result = spawn.sync(
     process.execPath,
     nodeArgs
-      .concat(require.resolve('../scripts/' + script))
+      .concat(require.resolve(`../scripts/${script === 'build-ssr' ? 'build' : script}`))
       .concat(args.slice(scriptIndex + 1)),
     { stdio: 'inherit' }
   );