+++ tasks/e2e-kitchensink-eject.sh
@@ -134,8 +134,8 @@ REACT_APP_SHELL_ENV_MESSAGE=fromtheshell \
   yarn build
 
 # Check for expected output
-exists build/*.html
-exists build/static/js/main.*.js
+exists build/web/*.html
+exists build/web/static/js/main.*.js
 
 # Unit tests
 REACT_APP_SHELL_ENV_MESSAGE=fromtheshell \